Fix compilation with Delta debug dumping enabled.
git-svn-id: http://svn.drobilla.net/lad/trunk/ingen@5005 a436a847-0d15-0410-975c-d299462d15a1

+++ src/server/events/Delta.cpp
@@ -71,7 +71,7 @@ Delta::Delta(Engine&           engine,
 	}
 
 #ifdef DUMP
-	std::cerr << "Delta " << subject << " : " << context << " {" << std::endl;
+	std::cerr << "Delta " << subject << " : " << (int)context << " {" << std::endl;
 	typedef Resource::Properties::const_iterator iterator;
 	for (iterator i = properties.begin(); i != properties.end(); ++i) {
 		std::cerr << "    + " << i->first
